refact: update delete and put endpoints

+++ StarWars.Characters/Application/Characters/RemoveCharacterCommand.cs
@@ -7,7 +7,7 @@ namespace StarWars.Characters.Application.Characters;
 using Command = RemoveCharacterCommand;
 using Result = OneOf<bool, RemoveCharacterError>;
 
-public record RemoveCharacterCommand(int Id) : IRequest<Result>, ITransactional;
+public record RemoveCharacterCommand(int Id, int UserId) : IRequest<Result>, ITransactional;
 
 internal class RemoveCharacterCommandHandler(ICharacterRepository characterRepository) : IRequestHandler<Command, Result> {
     public async Task<Result> Handle(Command cmd, CancellationToken c) => await RemoveCharacterAsync(cmd, c);
@@ -16,6 +16,7 @@ internal class RemoveCharacterCommandHandler(ICharacterRepository characterRepos
         var character = await characterRepository.GetByIdOrDefaultAsync(cmd.Id, c);
         
         if (character == null) return RemoveCharacterError.CharacterNotFound;
+        if (character.CreatedBy.Id != cmd.UserId) return RemoveCharacterError.OnlyAuthorCanRemoveCharacter;
 
         characterRepository.Remove(character);
         
@@ -25,4 +26,5 @@ internal class RemoveCharacterCommandHandler(ICharacterRepository characterRepos
 
 public enum RemoveCharacterError {
     CharacterNotFound,
+    OnlyAuthorCanRemoveCharacter,
 }
